+++ client/src/commonMain/kotlin/eu/vendeli/rethis/providers/PooledConnectionProvider.kt
@@ -16,7 +16,8 @@ internal class PooledConnectionProvider(
         ConnectionPool(node.socket, cfg, connectionFactory, rootJob)
     }
 
-    override suspend fun execute(request: CommandRequest): Buffer = withConnection { it.doRequest(request.data) }
+    override suspend fun execute(request: CommandRequest): Buffer =
+        withConnection { it.doRequest(request.data, commandTimeout = client.cfg.commandTimeout) }
 
     override fun close() {
         if (client.cfg.pool.closeGracefully) coRunBlocking { pool.closeGracefully() }