+++ client/src/commonMain/kotlin/eu/vendeli/rethis/providers/SingleConnectionProvider.kt
@@ -12,7 +12,8 @@ internal class SingleConnectionProvider(
     override val node: Address,
     private val client: ReThis,
 ) : ConnectionProvider() {
-    override suspend fun execute(request: CommandRequest): Buffer = withConnection { it.doRequest(request.data) }
+    override suspend fun execute(request: CommandRequest): Buffer =
+        withConnection { it.doRequest(request.data, commandTimeout = client.cfg.commandTimeout) }
 
     override fun close() {}